FIX(server,client): Validate the correct certifiacte from a clients chain
Currently, obtaining and validating the clients certificate chain is a mess. While obtaining the chain (which is ordered, starting with the peer's immediate certificate and ending with the CA's certificate), the peers immediate certificate is added again, as last certificate. Then, while validating, the last certificate is checked. This approach works (since the validated certificate is the one that is expliticly added), but puts the whole concept of a ordered certificate chain to absurdity.

This commit fixes that. First, the chain it is returned unaltered in its original form (ordered and starting with the peer's immediate certificate and ending with the CA's certificate). Then, while validating, the first certificate in this chain is checked.

The client uses the hash of the server's certificate to store some information about the server in the client's database. Since now the unaltered chain is returned, the first certificate in the chain must be examined instead of the last.

Fixes #3523 (partially)