FEAT(server): Make tolerated server ssl errors configurable
If a client makes a connection to a server and presents a ssl certificate, ssl errors can arise, e.g. because of an expired client certificate. Up till now, many of these errors were just ignored. This commit makes it possible to configure which ssl errors a server should allow to happen and on which a client should be force-disconnected.

The allowed ssl errors are parsed from the servers .ini and transformed to QSslError:SslError by comparison using a map. This matching is case insenitive. The maps that are used for matching the names to errors and vice versa are given in Meta::getSslNameErrorMap() and Meta::getSslErrorNameMap().

The new ini-option is named "allowedClientSslErrors" and used as described below:
allowedClientSslErrors (default: AllowNoPeerCertificate, AllowSelfSignedCertificate, AllowSelfSignedCertificateInChain, AllowUnableToGetLocalIssuerCertificate, AllowUnableToVerifyFirstCertificate, AllowHostNameMismatch, AllowCertificateNotYetValid, AllowCertificateExpired)
If a client presents a certificate to the server, the certificate is validated according to Murmurs Client Truststore (MCTS), which by default contains murmurs own chain-of-trust as well as the hostmachines CAs. Errors might arrise during validation of the clients certificate, but some are not critical and thus ignored. A server admin can configure, what errors should ignored using a comma separated list of the following options: AllowUnableToGetIssuerCertificate, AllowUnableToDecryptCertificateSignature, AllowUnableToDecodeIssuerPublicKey, AllowCertificateSignatureFailed, AllowCertificateNotYetValid, AllowCertificateExpired, AllowInvalidNotBeforeField, AllowInvalidNotAfterField, AllowSelfSignedCertificate, AllowSelfSignedCertificateInChain, AllowUnableToGetLocalIssuerCertificate, AllowUnableToVerifyFirstCertificate, AllowCertificateRevoked, AllowInvalidCaCertificate, AllowPathLengthExceeded, AllowCertificateUntrusted, AllowCertificateRejected, AllowSubjectIssuerMismatch, AllowAuthorityIssuerSerialNumberMismatch, AllowNoPeerCertificate, AllowHostNameMismatch, AllowUnspecifiedError, AllowNoSslSupport, AllowCertificateBlacklisted. If the server was compiled using a Qt version greater or equal 5.13.0, these additional values can be used: AllowCertificateStatusUnknown, AllowOcspNoResponseFound, AllowOcspMalformedRequest, AllowOcspMalformedResponse, AllowOcspInternalError, AllowOcspTryLater, AllowOcspSigRequred, AllowOcspUnauthorized, AllowOcspResponseCannotBeTrusted, AllowOcspResponseCertIdUnknown, AllowOcspResponseExpired, AllowOcspStatusUnknown. If no errors should be allowed, an empty list can be provided, or the value "None". An exception to the above rule are "Invalid Purpose" ssl errors, which are not configurable and always allowed.

The default values where choosen this way to match the current behaviour.

Impelments #3523 (partially)