FEAT(server): Make the trust store of the server configurable
Currently, if a client makes a connection to a server and provides a ssl certificate, this certificate is validated and errors are treated according to the "allowedClientSslErrors" ini-option. This commit makes it configurable, which client certificates should be trusted in this validation process. The general idea is to introduce the Murmur Client Truststore (MCTS), which contains all the certificate authorities to trust.

The following ini-options were added:
mctsAddOwnCertificateAsCA (default: true)
Adds murmurs own certificate (without the chain) as CA to the MCTS.

mctsAddOwnCAs (default: true)
Adds the chain of murmurs own certificate (but not the its own certificate) as CAs to the MCTS.

mctsHostsCAs (default: true)
If this is set to true, the CAs of the hostmachine are added to the MCTS.

mctsExtraCAs (default: empty)
Path to a pem file containing one to multiple certificates to add to the MCTS.

The default values where choosen this way to match the current behaviour.

Impelments #3523 (partially)