CHANGE(Server): Only send version messages as response
Up to this commit, a server sended a version message right in the encrypt signal of the socket. Now, the server only responds to a received version message from a client by sending an own version message.

The purpose behind this is to guarantee that a server never sends the first message in the communication with a client.