FEAT(Server): Websocket support
This commit adds websocket support.

Reminder: A message in the mumble control message protocol is defined as a two byte message type and a four byte message length, followed by the actual message data. For simplicity this protocol will be refered to as the "raw protocol" in the following.

Previous commits guaranteed that the first message in the communication between a server and a client is always sended from the client. In the websocket protocol, this is mandatory, as the client starts by sending a handshake to the server. This handshake always starts with the ASCII letters GE. Since upon reading the first message, the server can not know, wheter to use the raw protocol or the websocket protocol, it will always assume the raw protocol. Consequently, the server will transfrom this first two bytes in the communication according to the raw protocol in a message type. The ASCII letters GE correspond to the numerical message type 18245. If this message type is read (as first in a communication), the server knows that the client wants to use the websocket protocol and inteprets the contents of the connection that way.

The websocket protocol knows concepts of "framing", message types and message lengths. The server ignores these. While this server implementation follows the websocket protocol for control frames (e.g. respond to ping frames with pong frames), non-control websocket frames are all treated as binary data. A client that wants to use the websocket protocol MUST still construct a message according to the raw protocol, and send this as binary websocket message to the server. Boldly spoken, the server will strip everything websocket related and just look at the binary data.